Fix second CI round: Gradle/AGP/Kotlin/compileSdk bumps + num.clamp() bug, restrict workflow to push-to-main only
- Bump AGP 8.6.0 -> 8.11.1, Kotlin 1.9.24 -> 2.2.20, Gradle 8.7 -> 8.14.0
  (wrapper properties + CI's wrapper-regeneration step): the hard
  CheckAarMetadataWorkAction failure required AGP >=8.9.1 + compileSdk 36
  for androidx.browser/androidx.core(-ktx) (transitive deps of app_links,
  shared_preferences_android, url_launcher_android); jumped to the higher
  numbers Flutter's own "will soon be dropped" warnings named, to avoid a
  third round over the same category.
- Bump compileSdk/targetSdk 35 -> 36 in app/build.gradle to match.
- Fix num.clamp() always returning num (never the double/int receiver's
  own type) at 4 call sites: profile_detail_screen.dart (rewritten
  without clamp), progress_bar.dart, stacked_bar_chart.dart,
  doughnut_chart.dart. This broke strict double/int call sites like
  LinearProgressIndicator.value and Canvas.drawArc's sweepAngle.
- Restrict flutter-build.yml to only run on push to main: drop the
  master branch and the pull_request trigger, keep workflow_dispatch as
  a manual escape hatch.

+++ flutter_app/lib/widgets/progress_bar.dart
@@ -13,7 +13,10 @@ class AppProgressBar extends StatelessWidget {
     return ClipRRect(
       borderRadius: BorderRadius.circular(height / 2),
       child: LinearProgressIndicator(
-        value: (value.clamp(0, 100)) / 100,
+        // num.clamp() always returns num, even on a double receiver — the
+        // .toDouble() is load-bearing, not decorative (LinearProgressIndicator
+        // .value is strictly double?).
+        value: value.clamp(0, 100).toDouble() / 100,
         minHeight: height,
         backgroundColor: colors.border,
         valueColor: AlwaysStoppedAnimation<Color>(colors.primary),
